refactor offline clean up
