+++ src/libs/Share/index.js
@@ -3,6 +3,7 @@ import Navigation from '../Navigation/Navigation';
 const dismiss = () => Navigation.dismissModal();
 
 export default {
+    cleanUpActions: () => [],
     dismiss,
     isShareExtension: false,
     registerListener: () => ({remove: () => {}}),
